+++ entrypoints/background/index.ts
@@ -1,0 +1,7 @@
+import { registerContextMenus } from "./shopkeeperContextMenus";
+
+export default defineBackground(() => {
+  browser.runtime.onInstalled.addListener(() => {
+    registerContextMenus();
+  });
+});